Moved read_monsters endpoints to dqm1_endpoint router from main.py
